feat: API upload/delete
+ Minor bug fix for expired tokens on web
